fix: relay start issues by applying a different pattern.
