--- packages/core-logger/index.ts
@@ -1,3 +1,0 @@
-export * from "./src/manager";
-export * from "./src/logger";
-export * from "./src/index";
