+++ packages/core-logger/src/index.ts
@@ -1,11 +1,3 @@
-import { LogManager } from "./manager";
-
-const plugin = {
-    pkg: require("../package.json"),
-    alias: "logManager",
-    async register() {
-        return new LogManager();
-    },
-};
-
-export { plugin };
+export * from "./manager";
+export * from "./logger";
+export * from "./plugin";
