+++ packages/core-logger/src/plugin.ts
@@ -1,0 +1,11 @@
+import { LogManager } from "./manager";
+
+const plugin = {
+    pkg: require("../package.json"),
+    alias: "logManager",
+    async register() {
+        return new LogManager();
+    },
+};
+
+export { plugin };
