+++ packages/core-logger/tsconfig.json
@@ -3,5 +3,5 @@
     "compilerOptions": {
         "outDir": "dist"
     },
-    "include": ["index.ts", "src/**/**.ts"]
+    "include": ["src/**/**.ts"]
 }
